requirements.txt
sqlalchemy
fastapi
schedule
ansible
uvicorn
plotly-express
pandas
pydantic
python-decouple